refactor: merge duplicate code from sandboxed/non-sandboxed renderer init
